add utils + fix logic in maplayer
